refactor: Update PWA and importmap paths
Corrected the path for manifest.json and sw.js in the index.html to be relative. Adjusted the importmap in index.html to align dependencies with common import order and to ensure react and react-dom are correctly mapped. Updated README.md to emphasize deleting the `docs` folder before building for a clean compilation.